requirements.txt
AllanTools==2019.9
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.7
filelock==3.12.4
fonttools==4.43.1
imageio==2.31.5
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pip==23.3.1
platformdirs==3.11.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.3
shapely==2.0.1
six==1.16.0
trimesh==3.23.5
tzdata==2023.3
virtualenv==20.24.6